refactor(tests): Merge EOF stack underflow tests (#1384)
This combines the cases of `test_all_opcodes_stack_underflow()`
to `test_all_opcodes_variadic_stack_underflow()`.
By the way, this also improves some details, e.g. avoids instruction
sequence as `RETURN STOP`.
